--- src/main/java/students/student_renat_galickij/lesson_2/level_1_inter/level_2_inter/Task_5.java
@@ -1,6 +1,0 @@
-package students.student_renat_galickij.lesson_2.level_1_inter.level_2_inter;
-
-public class Task_5 {
-    public static void main(String[] args) {
-    }
-}
