+++ src/main/java/students/student_renat_galickij/lesson_2/level_2_inter/Task_4.java
@@ -1,0 +1,7 @@
+package students.student_renat_galickij.lesson_2.level_2_inter;
+
+public class Task_4 {
+    public static void main(String[] args) {
+
+    }
+}
