+++ src/main/java/students/student_renat_galickij/lesson_2/level_2_inter/Task_6.java
@@ -1,7 +1,9 @@
-package students.student_renat_galickij.lesson_2.level_1_inter.level_2_inter;
+package students.student_renat_galickij.lesson_2.level_2_inter;
 
-public class Task_7 {
+public class Task_6 {
+
         public static void main(String[] args) {
             System.out.println("You can login!");
         }
-}
+    }
+
