+++ src/main/java/students/student_renat_galickij/lesson_2/level_2_inter/Task_8.java
@@ -1,4 +1,4 @@
-package students.student_renat_galickij.lesson_2.level_1_inter.level_2_inter;
+package students.student_renat_galickij.lesson_2.level_2_inter;
 import java.util.Scanner;
 
 public class Task_8 {
