+++ src/main/java/students/student_renat_galickij/lesson_2/level_3_junior/Task_10.java
@@ -1,10 +1,10 @@
-package students.student_renat_galickij.lesson_2.level_1_inter.level_3_junior;
+package students.student_renat_galickij.lesson_2.level_3_junior;
 
 import java.util.Scanner;
 
 public class Task_10 {
     public static void main(String[] args) {
-        double per = 0, area = 0;
+        double per, area;
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter circle's radius");
         double radius = scanner.nextDouble();
