+++ src/main/java/students/student_renat_galickij/lesson_2/level_5_middle/Task_18.java
@@ -1,4 +1,4 @@
-package students.student_renat_galickij.lesson_2.level_1_inter.level_5_middle;
+package students.student_renat_galickij.lesson_2.level_5_middle;
 
 public class Task_18 {
         public static void main(String[] args) {
