+++ src/main/java/students/student_renat_galickij/lesson_2/level_6_middle/Task_24.java
@@ -1,4 +1,4 @@
-package students.student_renat_galickij.lesson_2.level_1_inter.level_6_middle;
+package students.student_renat_galickij.lesson_2.level_6_middle;
 
 public class Task_24 {
     public static void main(String[] args) {
